Reorganize shared resources by feature.
