+++ BwMelder.Core/RaceService.cs
@@ -1,5 +1,4 @@
-﻿using BwMelder.Shared.Dto;
-using BwMelder.Shared.Services;
+﻿using BwMelder.Shared.Races;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
