+++ BwMelder.Shared/ClubCoaches/IClubCoachService.cs
@@ -1,11 +1,10 @@
-﻿using BwMelder.Shared.Dto;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BwMelder.Shared.Services;
+namespace BwMelder.Shared.ClubCoaches;
 
 public interface IClubCoachService
 {
