--- BwMelder.Shared/Dto/UserLoginRequest.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace BwMelder.Shared.Dto;
-
-/// <summary>
-/// DTO representing the credentials a user provided to login.
-/// </summary>
-public class UserLoginRequest
-{
-    public required string Username { get; set; }
-    public required string Password { get; set; }
-}
