+++ BwMelder.Shared/Nomination/INominationService.cs
@@ -1,12 +1,11 @@
 ﻿using BwMelder.Shared.Crews;
-using BwMelder.Shared.Dto;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BwMelder.Shared.Services;
+namespace BwMelder.Shared.Nomination;
 
 /// <summary>
 /// Describes a service to process nominations.
