+++ BwMelder.Shared/Nomination/NominateNewClubRequest.cs
@@ -1,4 +1,5 @@
-﻿using BwMelder.Shared.Utilities;
+﻿using BwMelder.Shared.Clubs;
+using BwMelder.Shared.Utilities;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
@@ -6,7 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BwMelder.Shared.Dto;
+namespace BwMelder.Shared.Nomination;
 
 public class NominateNewClubRequest
 {
