+++ BwMelder.Shared/Races/IRaceService.cs
@@ -1,11 +1,10 @@
-﻿using BwMelder.Shared.Dto;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace BwMelder.Shared.Services;
+namespace BwMelder.Shared.Races;
 
 /// <summary>
 /// Represents a service to manage races.
