+++ BwMelder/Authentication/AuthenticationStateExtensions.cs
@@ -1,5 +1,4 @@
-﻿using BwMelder.Shared.Dto;
-using Microsoft.AspNetCore.Components.Authorization;
+﻿using Microsoft.AspNetCore.Components.Authorization;
 using System.Security.Claims;
 
 namespace BwMelder.Authentication;
