+++ BwMelder/Components/Pages/Admin/Crews/NominateCrew.razor
@@ -1,7 +1,10 @@
 ﻿@page "/admin/crews/nominate"
 @rendermode InteractiveServer
 
+@using BwMelder.Shared.Clubs
 @using BwMelder.Shared.Crews
+@using BwMelder.Shared.Nomination
+@using BwMelder.Shared.Races
 
 @inject IClubService ClubService
 @inject INominationService NominationService
