+++ BwMelder/Components/Pages/Admin/Login.razor
@@ -4,6 +4,7 @@
 @attribute [AllowAnonymous]
 
 @using BwMelder.Authentication;
+@using BwMelder.Shared.Authentication;
 
 @inject AuthenticationHandler AuthHandler;
 @inject NavigationManager Navigation;
