+++ BwMelder/Components/UI/ClubCoachForm.razor
@@ -1,4 +1,6 @@
-﻿<EditForm Model="ClubCoach" FormName="ClubCoachForm" OnValidSubmit="OnValidSubmit">
+﻿@using BwMelder.Shared.ClubCoaches
+
+<EditForm Model="ClubCoach" FormName="ClubCoachForm" OnValidSubmit="OnValidSubmit">
     <ObjectGraphDataAnnotationsValidator />
     <ValidationSummary />
     <div class="mb-3">
