+++ BwMelder/Components/UI/ContactSubform.razor
@@ -1,4 +1,6 @@
-﻿@* Contact is nested in different types. *@
+﻿@using BwMelder.Shared.Common
+
+@* Contact is nested in different types. *@
 @inherits Editor<ContactRequest>
 
 <div class="mb-3">
