+++ BwMelder/Components/_Imports.razor
@@ -12,5 +12,4 @@
 @using BwMelder.Components
 @using BwMelder.Components.UI
 @using BwMelder.DtoExtensions
-@using BwMelder.Shared.Dto
-@using BwMelder.Shared.Services
+@using BwMelder.Shared.Common
